+++ src/addons/character_dna/release_notes.md
@@ -2,6 +2,8 @@
 
 * Made DNA component type detection more robust [#337](https://github.com/poly-hammer/character-dna-addon/issues/337)
 * Fixed body animation pose mode switching bug
+* Fixed view options panel lagging draw calls
+* Fixed operator for adding [Texture Logic Node](https://docs.polyhammer.com/character-dna-addon/terminology/#texture-logic) now under `Add > Rig Logic > Head Texture Logic` in material graph
 
 ## Tests Passing On